+++ app/assets/javascripts/application.js
@@ -14,5 +14,4 @@
 //
 //= require jquery
 //= require jquery_ujs
-//= require turbolinks
 //= require_tree .
